server/requirements.txt
absl-py==0.12.0
anytree==2.8.0
appdirs==1.4.4
argon2-cffi==20.1.0
asn1crypto==1.4.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
Brlapi==0.8.2
bs4==0.0.1
btrfsutil==5.11.1
CacheControl==0.12.6
cachetools==4.2.2
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==3.4.7
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.1
distro==1.5.0
dnspython==1.16.0
entrypoints==0.3
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12.0
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.1
importlib-metadata==4.0.1
incremental==21.3.0
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isc==2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
libfdt==1.6.0
libmsym==0.2.4
lockfile==0.12.2
louis==3.17.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mistune==0.8.4
more-itertools==8.7.0
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netsnmp-python==1.0a1
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.0
OpenCC==1.1.2
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pep517==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
ply==3.11
prettytable==2.0.0
progress==1.5
prometheus-client==0.10.1
prompt-toolkit==1.0.14
protobuf==3.15.8
ptyprocess==0.7.0
pulsemixer==1.5.1
pwquality==1.4.4
pyalpm==0.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycparser==2.20
pycuda==2020.1
Pygments==2.8.1
PyHamcrest==1.10.0
PyInquirer==1.0.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.8.1
pyrsistent==0.17.3
pyte==0.8.0
python-dateutil==2.8.1
python-libtorrent==1.2.11
pytoml==0.1.21
pytools==2021.2.6
PyYAML==5.4.1
pyzmq==22.0.3
QOpenVPN==2.0.0
qtconsole==5.1.0
QtPy==1.9.0
rados==2.0.0
rbd==2.0.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
resolvelib==0.5.4
retrying==1.3.3
rgw==2.0.0
rsa==4.7.2
scipy==1.6.3
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
SSSDConfig==2.4.2
team==1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0rc2
tensorflow-estimator==2.5.0rc0
tensorflow-gpu @ file:///build/tensorflow/src/tmpcuda/tensorflow_gpu-2.4.1-cp39-cp39-linux_x86_64.whl
tensorflow-hub==0.7.0
tensorflowjs==2.0.1
termcolor==1.1.0
terminado==0.9.4
termtosvg==1.1.0
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
youtube-dl==2020.12.5
zipp==3.4.1
zope.interface==5.4.0
